basic tests to check status and project id
